refactor(db): move archived identity store ownership
